지난 전시 상세보기 틀 - YG

+++ app/src/main/java/com/yg/yourexhibit/Retrofit/RetrofitGet/ExhibitGoingResult.java
@@ -16,6 +16,6 @@ public class ExhibitGoingResult {
     int start_date;
     int end_date;
     String exhibition_picture;
-    int avg;
+    double avg;
     String flag;
 }
